Bot/requirements.txt
anyio==4.6.2.post1
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
numpy==2.1.3
pandas==2.2.3
playsound==1.2.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
